+++ src/components/ConstructorPage/ConstructorPage.jsx
@@ -2,7 +2,7 @@ import BurgerConstructor from "../BurgerConstructor/BurgerConstructor";
 import BurgerIngredients from "../BurgerIngredients/BurgerIngredients";
 import Wrapper from "../Wrapper/Wrapper";
 import PropTypes from 'prop-types';
-import PropTypesBurger from "../utils/PropTypesShape";
+import PropTypesBurger from "../../utils/PropTypesShape";
 
 
 function ConstructorPage() {
