+++ src/components/cardList/CardList.jsx
@@ -1,7 +1,7 @@
 import Card from "../Card/Card";
 import Style from './CardList.module.css';
 import PropTypes from 'prop-types';
-import PropTypesBurger from '../utils/PropTypesShape';
+import PropTypesBurger from '../../utils/PropTypesShape';
 
 function CardList(props) {
   const { category, items } = props;
